resume_pipeline/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pymysql
requests
python-dotenv
pydantic-settings
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pytesseract
easyocr
pdfminer.six
pdf2image
pdfplumber
python-magic
tqdm
faiss-cpu
jsonschema
